requirements.txt
asgiref==3.2.7
bootstrap==4.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-bootstrap-pagination==1.7.1
django-ckeditor==5.9.0
django-embed-video==1.3.2
django-js-asset==1.2.2
django-social-share==1.4.0
idna==2.9
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
django-smart-selects==1.5.6
django-cities-light
django-paypal
django-crispy-forms==1.9.1
django-countries==6.1.2
django-intl-tel-input==0.3.1